+++ projects/rocthrust/thrust/system/hip/detail/copy_if.h
@@ -56,7 +56,7 @@
 #  include <thrust/system/hip/detail/par_to_seq.h>
 #  include <thrust/system/hip/detail/util.h>
 
-#  include <cstdint>
+#  include _THRUST_STD_INCLUDE(cstdint)
 
 THRUST_NAMESPACE_BEGIN
 // XXX declare generic copy_if interface