+++ projects/rocthrust/thrust/system/hip/detail/reduce_by_key.h
@@ -59,7 +59,7 @@
 #    include _THRUST_STD_INCLUDE(iterator)
 #  endif
 
-#  include <cstdint>
+#  include _THRUST_STD_INCLUDE(cstdint)
 
 // rocprim include
 #  include <rocprim/rocprim.hpp>